smef/requirements.txt
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
checksumdir==1.2.0
cleo==0.8.1
clikit==0.6.2
coloredlogs==15.0.1
crashtest==0.3.1
cryptography==37.0.4
cycler==0.11.0
distlib==0.3.5
filehash==0.2.dev1
filelock==3.7.1
fonttools==4.33.3
html5lib==1.1
humanfriendly==9.2
idna==3.3
importlib-metadata==4.12.0
jeepney==0.8.0
keyring==23.7.0
kiwisolver==1.4.2
lockfile==0.12.2
matplotlib==3.5.2
msgpack==1.0.4
numpy==1.21.2
packaging==20.9
pandas==1.3.2
pastel==0.2.1
pexpect==4.8.0
Pillow==9.1.0
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.1.14
poetry-core==1.0.8
ptyprocess==0.7.0
pycparser==2.21
pylev==1.4.0
pyparsing==3.0.8
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.4
python-dateutil==2.8.2
pytz==2021.1
QDarkStyle==3.0.2
qtmodern==0.2.0
QtPy==1.10.0
requests==2.28.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.2
shellingham==1.4.0
six==1.16.0
smef==1.0
tomlkit==0.11.1
urllib3==1.26.11
virtualenv==20.16.1
webencodings==0.5.1
zipp==3.8.1
